pyproject.toml
[project]
name = "do-it-agent"
version = "0.1.0"
description = "Voice-triggered meeting agent: Calendar + Gmail actions via MeetStream webhooks and Scalekit"
readme = "DO-IT-AGENT.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "python-dotenv>=1.1.1",
    "scalekit-sdk-python>=2.6.1",
    "openai>=1.0.0",
    "requests>=2.31.0",
]
